Reorganised codes to individual files
